# apps/layer-gateway/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hevlayer API
  description: 'Hevlayer-specific endpoints that extend the Turbopuffer API. These
    endpoints are not part of the Turbopuffer client — they are added by the Hevlayer
    proxy layer for document retrieval (cache), document cache operations, snapshot
    jobs, scans, result counts, and namespace metadata.

'
  version: 0.1.0
  contact:
    name: Hev
    url: https://github.com/hev/hevlayer
servers:
- url: http://localhost:8080
  description: Local development
- url: http://hevlayer.internal:8080
  description: In-cluster (Kubernetes)
security:
- bearerAuth: []
paths:
...
  "/v1/namespaces":
    get:
      operationId: listTurbopufferNamespaces
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: GET /v1/namespaces
      summary: List Turbopuffer namespaces
      description: Raw Turbopuffer-compatible namespace listing passthrough.
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: prefix
        in: query
        required: false
        schema:
          type: string
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 1000
      responses:
        '200':
          description: Namespaces listed in Turbopuffer's native shape
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TurbopufferNamespaceList"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/v1/namespaces/{namespace}/schema":
    get:
      operationId: getTurbopufferNamespaceSchema
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: GET /v1/namespaces/{namespace}/schema
      summary: Get Turbopuffer namespace schema
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      responses:
        '200':
          description: Namespace schema
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TurbopufferSchema"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    post:
      operationId: updateTurbopufferNamespaceSchema
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: POST /v1/namespaces/{namespace}/schema
      summary: Update Turbopuffer namespace schema
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TurbopufferSchema"
      responses:
        '200':
          description: Updated namespace schema
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TurbopufferSchema"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/v1/namespaces/{namespace}/metadata":
    get:
      operationId: getTurbopufferV1NamespaceMetadata
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: GET /v1/namespaces/{namespace}/metadata
      summary: Get Turbopuffer v1 namespace metadata
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      responses:
        '200':
          description: Namespace metadata
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceMetadata"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    patch:
      operationId: updateTurbopufferNamespaceMetadata
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: PATCH /v1/namespaces/{namespace}/metadata
      summary: Update Turbopuffer namespace metadata
      description: Raw passthrough for upstream metadata configuration, currently
        used for namespace pinning.
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TurbopufferMetadataPatch"
      responses:
        '200':
          description: Updated namespace metadata
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceMetadata"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
...
  "/v2/namespaces/{namespace}":
    post:
      operationId: writeNamespace
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: POST /v2/namespaces/{namespace}
      x-hevlayer-client-aliases:
      - operationId: branchNamespace
        upstreamKey: POST /v2/namespaces/{namespace}?stainless_overload=branchFrom
        query:
          stainless_overload: branchFrom
        requestBodySchema:
          "$ref": "#/components/schemas/TurbopufferBranchFromRequest"
        responseSchema:
          "$ref": "#/components/schemas/TurbopufferWriteResponse"
      - operationId: copyNamespace
        upstreamKey: POST /v2/namespaces/{namespace}?stainless_overload=copyFrom
        query:
          stainless_overload: copyFrom
        requestBodySchema:
          "$ref": "#/components/schemas/TurbopufferCopyFromRequest"
        responseSchema:
          "$ref": "#/components/schemas/TurbopufferWriteResponse"
      summary: Write namespace rows
      description: 'Native Turbopuffer write body for upserts, deletes, patches, filter
        writes, schema updates, branch, copy, encryption, and related write options.
        Row-producing writes are server-stamped and mirrored into Layer''s document
        cache best-effort before the upstream call when ids are explicit. Conditional
        and filter writes request affected ids from Turbopuffer and apply cache invalidation
        and write-triggered UDF enqueue from the upstream affected-id set after success.

'
      tags:
      - namespaces
      - documents
      parameters:
      - "$ref": "#/components/parameters/namespace"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TurbopufferWriteRequest"
      responses:
        '200':
          description: Native Turbopuffer write response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TurbopufferWriteResponse"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '502':
          description: Turbopuffer write failed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    delete:
      operationId: deleteNamespace
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: DELETE /v2/namespaces/{namespace}
      summary: Delete a namespace
      description: 'Hard-delete the upstream Turbopuffer namespace, then purge gateway-owned
        namespace state: Aerospike document rows and snapshot mirrors, S3 snapshots/search
        history/clickstream/shard metadata, in-memory job/cache/consistency state,
        and the operator-discovered Index CR when Index garbage collection is enabled.
        Turbopuffer 404 is treated as idempotent success and still triggers local
        cleanup. Other Turbopuffer failures return 502 before local cleanup runs.
        Local cleanup failures after the upstream delete return 502 so callers can
        retry the same delete.

'
      tags:
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      responses:
        '200':
          description: Namespace deleted and local cleanup completed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StatusResponse"
        '502':
          description: Turbopuffer delete failed or local cleanup was incomplete
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
...
  "/v2/namespaces":
    get:
      operationId: listNamespaces
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      summary: List namespaces
      description: List namespaces with optional prefix filter and cursor pagination.
      tags:
      - namespaces
      parameters:
      - name: prefix
        in: query
        required: false
        schema:
          type: string
        description: Filter namespaces by prefix
      - name: cursor
        in: query
        required: false
        schema:
          type: string
        description: Pagination cursor from a previous response
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
      responses:
        '200':
          description: Namespaces listed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceList"
        '502':
          description: Upstream error fetching the namespace list
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/v2/namespaces/{namespace}/metadata":
    get:
      operationId: getNamespaceMetadata
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: GET /v2/namespaces/{namespace}/metadata
      summary: Get namespace metadata
      description: 'Get metadata for a namespace including schema, row counts, byte
        sizes, and custom config.

'
      tags:
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      responses:
        '200':
          description: Namespace metadata
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceMetadata"
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
...
components:
  parameters:
    namespace:
      name: namespace
      in: path
      required: true
      schema:
        type: string
      description: Namespace name
...
  schemas:
...
    TurbopufferNamespaceList:
      type: object
      required:
      - namespaces
      properties:
        namespaces:
          type: array
          items:
            "$ref": "#/components/schemas/TurbopufferNamespaceSummary"
        next_cursor:
          type: string
    TurbopufferSchema:
      type: object
      additionalProperties: true
    TurbopufferMetadataPatch:
      type: object
      additionalProperties: true
      properties:
        pinning: {}
    TurbopufferWriteRequest:
      type: object
      additionalProperties: true
      description: Native Turbopuffer write body (`upsert_rows`, `patch_rows`, `deletes`,
        filter writes, copy, branch, schema, encryption, etc.).
    TurbopufferBranchFromRequest:
      type: object
      additionalProperties: true
      properties:
        branch_from_namespace:
          type: object
          additionalProperties: true
      required:
      - branch_from_namespace
    TurbopufferCopyFromRequest:
      type: object
      additionalProperties: true
      properties:
        copy_from_namespace:
          oneOf:
          - type: string
          - type: object
            additionalProperties: true
      required:
      - copy_from_namespace
...
    NamespaceMetadata:
      type: object
      required:
      - id
      - schema
      - approx_logical_bytes
      - approx_row_count
      - created_at
      - updated_at
      properties:
        id:
          type: string
        schema:
          type: object
          additionalProperties: true
        approx_logical_bytes:
          type: integer
        approx_row_count:
          type: integer
        created_at:
          type: string
          format: date-time
        last_write_at:
          type:
          - string
          - 'null'
          format: date-time
        updated_at:
          type: string
          format: date-time
        config:
          type: object
          additionalProperties: true
        index:
          "$ref": "#/components/schemas/IndexState"
        layer:
          "$ref": "#/components/schemas/NamespaceMetadataLayer"
...
    Error:
      type: object
      required:
      - error
      - message
      properties:
        error:
          type: string
        message:
          type: string
...